Partial TypeScript refactors
