+++ src/consts/colors.mts
@@ -1,4 +1,4 @@
-const colors = {
+const colors: { [key: string]: string } = {
     Primary:   "b068a8",
     Secondary: "d9b69e",
     Error:     "db4b4b",
