--- src/consts/emojis.mjs
@@ -1,15 +1,0 @@
-const emojis = {
-  like: "<:like:1264163067655229510>",
-  upvote: "<:upvote:1264163489073860620>",
-  downvote: "<:downvote:1264163504173224006>",
-
-  booster:    "<:booster:1264162949753212928>",
-  supporter:  "<:supporter:1264162874465456200>",
-  translator: "<:translator:1335008128323944528>",
-  portalmod:  "<:portalmod:1264163032876060756>",
-  home:       "<:phantys_home:1264163401962229780>",
-
-  yo: "<:yo:1349437436719005749>"
-}
-
-export default emojis;
